feat: admin controlled descriptions and ownership
few cosmetic changes, db Book model modified, onSave script defined
