+++ static/scripts/onSave.js
@@ -1,0 +1,12 @@
+document.getElementById('submitButtonIcon').addEventListener('click', function() {
+    document.getElementById('descriptions').submit();
+});
+document.getElementById('submitButtonIcon').addEventListener('click', function() {
+    document.getElementById('owned').submit();
+});
+document.getElementById('submitButtonText').addEventListener('click', function() {
+    document.getElementById('descriptions').submit();
+});
+document.getElementById('submitButtonText').addEventListener('click', function() {
+    document.getElementById('owned').submit();
+});
