+++ static/styles/tables.css
@@ -1,3 +1,7 @@
+.table-img {
+    width: 140px;
+}
+
 .table-text {
     color: rgba(255, 255, 255, 0.6);
     font-family: 'Cambria';
@@ -6,6 +10,13 @@
 
 .table-header{
     background-color: #292929;
+}
+
+.table-header-title {
+    color: white !important;
+    font-size: 20px;
+    font-family: 'Cambria';
+    width: 100vw;
 }
 
 .table-header-text {
@@ -34,4 +45,5 @@
 
 .table-hover tbody tr:hover td, .table-hover tbody tr:hover th {
     color: white;
+    font-weight: bold;
 }
