+++ templates/booklist.html
@@ -7,22 +7,34 @@
         <p>Pełna lista Kunderów:</p>
     </div>
     <div class="rel">
+      <form method="POST" id="owned" name="owned">
         <table class="table table-striped table-bordered table-hover">
             <thead class="table-header">
               <tr>
-                <th scope="col"><b class="table-header-text">Czy Natka ma?</b></th>
-                <th scope="col"><b class="table-header-text">Tytuł</b></th>
+                <th scope="col" class="table-header-text">Czy Natka ma?</th>
+                <th scope="col" class="table-header-title">Tytuł</th>
+                <th scope="col" class="table-img"></th>
               </tr>
             </thead>
             <tbody>
             {% for book in books %}
             <tr>
-                <td class="table-text">Tak</td>
+                <td class="table-text">
+                  <div>
+                    <select id="{{ book.id }}" name="{{ book.id }}" class="" {% if not session.get("logged", False) %} disabled {% endif %}>
+                      <option selected value="{% if book.check == 'yes' %}yes{% else %}no{% endif %}">{% if book.check == "yes" %} Tak {% else %} Nie {% endif %}</option>
+                      <option value="{% if book.check == 'yes' %}no{% else %}yes{% endif %}">{% if book.check == "yes" %} Nie {% else %} Tak {% endif %}</option>
+                      <option value="loan">Wypożyczona</option>
+                    </select>
+                  </div>
+                </td>
                 <td class="table-text">{{ book.name }}</td>
+                <td class=""><img src="{{ static_link(book.img) }}" alt="{{ book.name }}" class="table-img"></td>
             </tr>
             {% endfor %}
             </tbody>
           </table>
+        </form>
     </div>
 </div>
